fix: Do not erase status fields on MarkNotSubscribed
If the observed Google Cloud resource is deleted before the source
object is deleted, these fields are nil while entering the finalizer. As
a result, all previously reconciled Pub/Sub resources get ignored
instead of being properly cleaned up.

Generally speaking, it is not useful to erase these fields. Status
conditions already convey that an error occurred, and leaving status
fields untouched is a good indicator that something was reconciled, at
some point (as opposed to a reconciliation that never succeeded).